requirements.txt
discord.py
python-dotenv
discord.py
flask
requests
